requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pydantic==2.9.0
pydantic[email]
Faker==30.0.0
python-dotenv==1.0.1
supabase==2.9.0
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
httpx
